FormCheckbox, FormCheckboxGroup: test moved into component folder
